app/ai_server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
chromadb
scikit-learn
numpy

